requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow
numpy
matplotlib
seaborn
requests
python-multipart
pydantic
